Renaming backup folder adding first unit tests
