+++ Pipfile
@@ -7,7 +7,6 @@ verify_ssl = true
 
 [packages]
 click = ">=3.3"
-mysql-connector = "==2.1.4"
 pid = ">=2.0"
 humanfriendly = ">=2.0"
 pytest = "*"
