+++ __init__.py
@@ -2,7 +2,7 @@
 
 
 from general_conf.generalops import GeneralClass
-from master_backup_script.backuper import Backup
+from backup_backup.backuper import Backup
 from general_conf.check_env import CheckEnv
 from prepare_env_test_mode.runner_test_mode import RunnerTestMode
 from .autoxtrabackup import *
