+++ backup_backup/__init__.py
@@ -1,0 +1,1 @@
+from backup_backup import backuper
