--- master_backup_script/__init__.py
@@ -1,1 +1,0 @@
-from master_backup_script import backuper
