+++ setup.py
@@ -11,7 +11,7 @@ if not isfile(path_config.config_path_file):
 setup(
     name='mysql-autoxtrabackup',
     version='2.0',
-    packages=['general_conf', 'backup_prepare', 'partial_recovery', 'master_backup_script',
+    packages=['general_conf', 'backup_prepare', 'partial_recovery', 'backup_backup',
               'prepare_env_test_mode', 'process_runner'],
     package_data={
         'prepare_env_test_mode': ['*.sh', '*.sql']
