+++ test/test_backup.bats
@@ -1,0 +1,12 @@
+#!/usr/bin/env bats
+
+# Run this BATS file to run Backup related tests
+
+DIRNAME=$BATS_TEST_DIRNAME
+
+@test "Running test_create_mysql_client_command" {
+run python -m pytest -vv ${DIRNAME}/test_backup.py::TestBackup::test_create_mysql_client_command
+  echo $output
+  [ $status -eq 0 ]
+}
+
